Revamped approach to symbols
Initially, the values of symbols were propogated up through the visitor,
only to be added to the assembly when absolutely needed. This was not
scalable in any way, so this new version simply pushes values to the
stack as soon as they are seen. This will make a lot of things way
easier